+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Features
+
+- *Experimental*: C#/.NET support (started in `2.0.0-beta.0`)
+  - Symbolicate managed stack frames on Android server-side ([#724](https://github.com/getsentry/sentry-godot/pull/724))
+
 ### Dependencies
 
 - Bump Sentry Android from v8.42.0 to v8.43.0 ([#731](https://github.com/getsentry/sentry-godot/pull/731))